remove-unnecessary-dir
